+++ internal/async/scheduler_task_config.go
@@ -6,6 +6,7 @@ import (
 	"github.com/hibiken/asynq"
 
 	"github.com/openkcm/cmk/internal/config"
+	"github.com/openkcm/cmk/utils/ptr"
 )
 
 // ScheduledTaskConfigProvider implements asynq PeriodicTaskConfigProvider interface.
@@ -19,16 +20,26 @@ func (p *ScheduledTaskConfigProvider) GetConfigs() ([]*asynq.PeriodicTaskConfig,
 	maps.Copy(tasks, config.PeriodicTasks)
 
 	for _, cfg := range p.Config.Scheduler.Tasks {
-		tasks[cfg.TaskType] = cfg
+		def := tasks[cfg.TaskType]
+		if cfg.Enabled != nil {
+			def.Enabled = cfg.Enabled
+		}
+		if cfg.Cronspec != "" {
+			def.Cronspec = cfg.Cronspec
+		}
+		if cfg.Retries != nil {
+			def.Retries = cfg.Retries
+		}
+		tasks[cfg.TaskType] = def
 	}
 
 	configs := make([]*asynq.PeriodicTaskConfig, 0)
 	for name, cfg := range tasks {
-		if !cfg.Enabled {
+		if cfg.Enabled != nil && !*cfg.Enabled {
 			continue
 		}
 
-		taskOpts := []asynq.Option{asynq.MaxRetry(cfg.Retries)}
+		taskOpts := []asynq.Option{asynq.MaxRetry(ptr.GetIntOrDefault(cfg.Retries, 0))}
 		if cfg.TimeOut > 0 {
 			taskOpts = append(taskOpts, asynq.Timeout(cfg.TimeOut))
 		}