+++ internal/config/config.go
@@ -170,17 +170,17 @@ func (s *Scheduler) GetTasks() map[string]Task {
 
 // Task holds a task config
 type Task struct {
-	Enabled    bool          `yaml:"enabled"`
+	Enabled    *bool         `yaml:"enabled"`
 	Cronspec   string        `yaml:"cronspec"`
 	TaskType   string        `yaml:"taskType"`
-	Retries    int           `yaml:"retries"`
+	Retries    *int          `yaml:"retries"`
 	TimeOut    time.Duration `yaml:"timeOut"`
 	FanOutTask FanOutTask    `yaml:"fanOutTask"`
 }
 
 type FanOutTask struct {
-	Enabled bool          `yaml:"enabled"`
-	Retries int           `yaml:"retries"`
+	Enabled *bool         `yaml:"enabled"`
+	Retries *int          `yaml:"retries"`
 	TimeOut time.Duration `yaml:"timeOut"`
 }